+++ src/assertions/index.ts
@@ -50,6 +50,7 @@ import { getNunjucksEngine } from '../util/templates';
 import { transform } from '../util/transform';
 import { AssertionsResult } from './AssertionsResult';
 import { handleBleuScore } from './bleu';
+import { handlePerplexityScore } from './perplexity';
 import { isSql } from './sql';
 import { getFinalTest, processFileReference } from './utils';
 import { validateXml, containsXml } from './xml';
@@ -1168,49 +1169,8 @@ ${
     };
   }
 
-  if (baseType === 'perplexity') {
-    if (!logProbs || logProbs.length === 0) {
-      throw new Error(
-        'Perplexity assertion does not support providers that do not return logProbs',
-      );
-    }
-    const sumLogProbs = logProbs.reduce((acc, logProb) => acc + logProb, 0);
-    const avgLogProb = sumLogProbs / logProbs.length;
-    const perplexity = Math.exp(-avgLogProb);
-
-    pass = assertion.threshold ? perplexity <= assertion.threshold : true;
-    return {
-      pass,
-      score: pass ? 1 : 0,
-      reason: pass
-        ? 'Assertion passed'
-        : `Perplexity ${perplexity.toFixed(2)} is greater than threshold ${assertion.threshold}`,
-      assertion,
-    };
-  }
-
-  if (baseType === 'perplexity-score') {
-    if (!logProbs || logProbs.length === 0) {
-      throw new Error(
-        'perplexity-score assertion does not support providers that do not return logProbs',
-      );
-    }
-    const sumLogProbs = logProbs.reduce((acc, logProb) => acc + logProb, 0);
-    const avgLogProb = sumLogProbs / logProbs.length;
-    const perplexity = Math.exp(-avgLogProb);
-    const perplexityNorm = 1 / (1 + perplexity);
-
-    pass = assertion.threshold ? perplexityNorm >= assertion.threshold : true;
-    return {
-      pass,
-      score: perplexityNorm,
-      reason: pass
-        ? 'Assertion passed'
-        : `Perplexity score ${perplexityNorm.toFixed(2)} is less than threshold ${
-            assertion.threshold
-          }`,
-      assertion,
-    };
+  if (['perplexity', 'perplexity-score'].includes(baseType)) {
+    return handlePerplexityScore(logProbs, assertion);
   }
 
   if (baseType === 'cost') {
